Inquiry into dispatchery based on interfaces.
Seems like this is syntactically kind of hard to access.  Possible, but not graceful.

Then again, given that we're *well* into use of reflect at the dispatch level, I guess it's not at all silly to say "make your funcs take the type param you want", and we use that.  (Except that doesn't work again for the pseudo-genericized funcs you'll end up holding if you're using an atlas-based handler.  So.. yeah, challenges.)

Signed-off-by: Eric Myhre <hash@exultant.us>