Attempt to replicate a json decoder and fit it to our interfaces.
(Lots of this code is riffing from golang stdlib; some snippets are also from the ugorji codecs.  Both can be considered under their respective OSS licenses.)

Probably *not* going to move forward with this exact set of stuff though.  On further reflection, there's...

- A: just plain quite a tangle here
- B: we want a continuation-passing-ready state machine, and only scanner is that (the rest leans pretty hard on the stack)
- C: following from B, we really want to revist our whole call style to *not* be pushing directly into the token sink calls; that seems like a weird stack arrangement to end up with.

So this code is mostly going into the dustbin, and we're gonna do a clean slate again here.

Signed-off-by: Eric Myhre <hash@exultant.us>