Snapshot of the reboot.
Driving the rebuild by starting with the decode function, because those are generally a bit wilder than the encodes -- might have much less type info in some cases.

Whole thing is based on the token sources and sinks *both* required to be state machines that handle continuations inside, so we can step them both in parallel with neither of them having a stack (and thus error handling privs) over the other.

Signed-off-by: Eric Myhre <hash@exultant.us>