Make the whole thing step for a wildcard interface being filled with a map.
Fixed wildcard step not behaving like enough of a machine to advance properly, and fixed token consts not really being right type (which makes them not consts?  which is dumb and sigh but whatever), and fixed some conditions in the top level step being obviously wrong as soon as a test was aimed at exercising them, and fixed some pointer things being not reffy enough.  Whew.

Signed-off-by: Eric Myhre <hash@exultant.us>