Attempt to defeat more possible over-optimizations by compiler.
I've seen these lines have effects in other benchmarks in the past, but they're not phasing anything here.

I'm *very* unclear on *how* so many of these are managing to be zero-alloc.  It doesn't look to me like the paths we're hitting in the `reflect` package should be zero-alloc at all.

Signed-off-by: Eric Myhre <hash@exultant.us>