Microbenchmarks of map iteration.
Using reflection carries a fairly high cost.  We can get about 16x improvements in tokenization throughput if simply using a compile-time-known key type!

So this -- for string keys at least, given that we're conceding that as the defacto Correct Type for Keys -- seems like a reasonable thing to do.  Other types... we'll see; this would likely be one of those alarming things that requires code generation, and I think there's a requirement for strong justification before embracing that cost.

Signed-off-by: Eric Myhre <hash@exultant.us>