Documentation for token machine design.
Specifically describe how this relates to automata theory.  The distinction between the drivers and the machine funcs is much easier to clarify when framed in terms of bounding complexity.

The names here are subject to change.  They very definitely do not happen to sort very well if you looked at this all in godoc right now.

Signed-off-by: Eric Myhre <hash@exultant.us>