Comment block on letting an atlas.Entry explicitly choose the machinery to handle a field value.
This is a feature that will probably be rarely used, but seems fundamentally correct to me: punting back all the way to the top level driver when recursing during an object walk (as the previous gen prototype did) strikes me as comparable to "Dependency Injection" in a very negative way.  Letting function calls pick which function calls they make is kind of a no-brainer.  Maybe for this codec job, the DI path is common for boilerplate saving, but explicit should always be king.

The biggest source of potential awkwardness with this feature is that the clearly correct thing to do is specify the machinery as, well, the actual FSMs.  Which are asymmetric on marshal and unmarshal.  I still can't get over how much you'd *think* would be symmetric about the two purposes for object walking, but really isn't.  Anyway, this might require folding the `atlas` package back into `obj` to avoid more cyclic import problems.

Signed-off-by: Eric Myhre <hash@exultant.us>