Add Suite to MarshalMachine function params.
Both `Reset` and `Step` -- sometimes you need new machinery and can't tell until you get to that value step; sometimes you can look it up right away and keep reusing the machine (slices and maps usually behave like the latter).

There are now two methods on Suite, depending on whether or not you have a pointer to a real thing, or just reflective type info.  I'm not a hundred percent confident that's going to be a useful distinction, but I'm going to keep it for now, and make further decisions on that after we see how complex it is to maintain both code paths vs the relevance in benchmarks once we get to a bigger system.

Signed-off-by: Eric Myhre <hash@exultant.us>