Add atlas.Field.AddrFunc test coverage for type info and readable/writable references.
Worth noting there's a conspicuous lack of error checking if one writes an AddrFunc that returns unexpected types of things.  This interface is very much for consenting adults who know what they're doing.  More training wheels may come later (but possibly not, because this *is* also the *fast* path and I'm loathe to compromise that).

Signed-off-by: Eric Myhre <hash@exultant.us>