In-progress snapshot of more marshal machines.
This is not fully functional, and the scope of changes has started to eclipse what I can reasonably try to fit in one commit and keep my sanity.  Testing is also an issue: I was hoping full tests of the marchal driver stack would be able to give good coverage for the individual FSMs, but that's turning out WAY too complex and needs to be addressed.

So, the next commit will be factoring down the marshal driver into an interface, so I can use a mocked version sans recursion in tests that target single MarshalMachine implementations.  This strikes me as dangerously close to "mocking", which is generally an indication you need to rethink your interfaces, but I'm fresh out of better ideas, short of changing the entire MarshalMachine.Step interface to optionally yield other MarshalMachines, which seems excessive.

Signed-off-by: Eric Myhre <hash@exultant.us>