Incrementally more correct and better-error-checked Grab method.
The "range" operating in TraverseToValue does very bad things if used on an uninitialized value: it silently does nothing, leaving you with the same value.  Is that... correct?  I guess.  But very definitely wrong, in the context of an atlas field?  Oh yes: it will leave you infinitely recursing (without actual recursion on the stack, mind, so it won't stop until you OOM).

Signed-off-by: Eric Myhre <hash@exultant.us>