Implement fieldname->fieldroute initialization for atlas and its entries.
Did indeed make the refect.Type info part of the atlas setup.  Trying to put it into the init method args and pushing it out that direction seems no better.  But yes, we now definitely have some redundancies again between specifying what type an atlas covers and what you want the demux to call for which types.  This isn't wrong, if you see atlases being used anywhere outside of that demux; just syntactically unfortunate for our primary use case... so something ought be done about that (then again, remember, we're still planning to make it a default behavior to automatically generate atlases and add them to the suite for new types, so *that's* really the primary use case, meaning maybe this won't be a major user-relevant issue at all)..

Signed-off-by: Eric Myhre <hash@exultant.us>