Alright: there.  Resolved the token mess for marshaller machines.
All tokens are refs to primitives again.  Writing test fixtures is ugly again as well, but this is apparently the least-ambiguous way to deal with things -- losing type information due to the way the test fixture tables are declared is really painful, and if this level of explicitness is the answer, fine.

AND WE'RE FINALLY BACK TO GREEN AGAIN.

Signed-off-by: Eric Myhre <hash@exultant.us>