Add marshal machine for pointers; struct atlas marshall machine tests now cover structs with (nil) pointers.
This will probably be worth a review for efficiency in the future, because it's allocating this whole machine to do the indirection step, and that seems both wasteful and likely to be a common occurence.  An alternative would be to make the whole `pick` function capable of dereferencing values... doing so would require a refactor (since it's currently guided only by type info), and potentially muddle the question of how much that function is really supposed to do, and how to test it, so... perhaps later.

Signed-off-by: Eric Myhre <hash@exultant.us>