Approx 1 metric ton of debug printf's.
I wish for some kind of utility that would let me journal all of these steps, then write tests asserting against them.  This would be *very* useful for making sure it's taking the correct, efficient steps though things, for example (e.g. type info is sufficient to jump us straight to the right machine without accidentally bouncing through an `interface{}` wildcard detangle step, and things such as that).  It could be done directly, just passing down a slice and so forth, but I'm loathe to write because it wouldn't be zero-cost to runtime performance (or, making it such would require increasingly alarming amounts of macro-like hacks, which, let's not).

Signed-off-by: Eric Myhre <hash@exultant.us>