Support and test coverage for marshal of recursive objects.
This requires our marshaller suite to work in terms of factory functions, so it can return new machines when "recursing" down our synthetic "stack".  All of our explicit routes for built-in types effectively already worked like this; most of the diff here is where a suite is set up to use atlas-based machines.

There's inefficiency for MarshalMachine implementations that are stateless (as literal machines are): the factory function gets called anyway.  I'm as yet unsure how to improve this situation, aside from resorting to a MarshalMachineFactoryOrStatelessMachine union type, which seems a bit naff.  We should be able to get a reasonable estimate of the overhead of the current approach by benchmarking `struct{w,x,y,z int}` vs `type T int; struct{w,x,y,z T}`.

How machine reuse to evade mallocs is either eased or hindered is also an entanglement in these choices.  But again, we'll see how big of an issue this is.  ISTM that in practice, when dealing with structs, one has minimal numbers of children, and caching machines doesn't matter too much because the $N$ is so small; it's when dealing with maps and slices that the fan-out ratio is quite high, and those are inherently easier to cache machines for because they know their type info up front and simply directly cache the correct value machine during their `Reset()`.

Signed-off-by: Eric Myhre <hash@exultant.us>