Benchmarks for wholescale conversion of object to json, so we can compare with stdlib.  Script to drive this.  Application of pprof to cpu use per func, for help in improving and refining each situation.
Results are already fascinating and useful.

For example!  For the int slice test, xlate is spending a whopping 11% of its overall time in 'runtime.convT2E' -- the function that takes care of boxing up a value into an interface.  Now, this is usually *not* a problem for our code, because we detected early on with our microbenchmarks that this follows a different (cheaper, malloc-free!) rule when you do it to *pointers*.  (I suppose it's documented somewhere as well, but se la vie.)  So why is this rearing its ugly head here?  Because of this: ` d.wr.Write([]byte(fmt.Sprintf("%d", *valp)))` -- our placeholder code for better terminal value serialization.  Derefering the 'valp' gives (as the name implies) the val without the pointer.  The `Sprintf` function then, because it takes `...interface{}` type arguments, boxes it.  And there's the pricetag.  This is no great issue, since those are indeed placeholders, and not very correct placeholders at that, but the information is facinating.

Signed-off-by: Eric Myhre <hash@exultant.us>