Experiment with avoiding alloc for literal machines entirely.
About another 5% disappears off the struct benchmarks.

Given that the combined number of struct machines and ptr machines hit by that benchmark is more than the number of literal leaves, we can expect the same or better bonuses if we preallocate slabs with all the possible needed machines.  I've been wary of this, because it starts to priviledge our code over any possible external interface implementers... but it seems the performance implications more or less force our hand.

Signed-off-by: Eric Myhre <hash@exultant.us>