Remove canaddr check in struct machine reset; was not correct.
This is a little embarassing; the combination of that check, and the error it then allocated (incorrectly; and also recursively embarassingly, not checked) was another 24% of our overall time.

We're now down to 4-digit ms per op on my hardware; same order of mag as the stdlib.  So, while I'm not keeping exact track, we're now down to slashing between 3~2.5x of our original overhead from fixing nickle-and-dime things.  Not bad.

Signed-off-by: Eric Myhre <hash@exultant.us>