First pass attempting to use allocfreetrace for ridiculously, amazingly high resolution insight into when allocations occur.
It's completely overwhelming.  Amazing, but overwhelming.  On the order of 1GB of logs generated for a bench of a single function.

It also doesn't work correctly as written.  `go test` ignores the count variable for benchmarks, irritatingly.  Also, the fraction of the log that comes from the test harnesses (and the ast parse...!!) is quite large.  But I'm going to swallow that irritation for now rather than dupe a bunch of code into another package with a custom main method and so on.

Signed-off-by: Eric Myhre <hash@exultant.us>