Touch to struct bench fixture explore behavior.
This delta increases our allocs by 6, and stdlib's by 1.

We're now at a whopping 45 allocs compared to stdlib's 4.  (Surprising we're doing this *well* with that much of a difference, really.)

So stdlib increased by... 1.  That's... a really large percentage, isn't it?  How could the step possibly be that big, when we're adding just one piece to a much more massive structure?

The 3 allocations stdlib previously did are:

- the encodingState struct, once.  This makes sense.  It's the rough equivalent of our MarshalDriver struct.
- a `json.encoderFunc`, once.  ONCE?!  We have FOUR different kinds of structures!  Hmm.
- a byte slice.  This is done by a bytes.Buffer growing, which happens to only come up once.

That single allocation of an encoderFunc is pretty fascinating.  The stack indicates it's from the very first object.  Evidentally all future uses are made invisible to us when looking through alloctrace microscope because escape analysis moves all of their impact to the stack.

The forth new allocation?  bytes.Buffer growing another time.  :/  Less exotic than one of my earlier theories (I mistakenly hypothesized that only types which were visited more than once triggered the encoderFunc alloc.)

This also shines a light on our side: a surprisingly large number of our allocs come from a cast to []byte in our json serializer, because we're *not* using bytes.WriteString's helpful internal buffering.  We'll want to fix that shortly.  (At some point, looking at the percent balance between the obj walker and serializer vs the same(ish) ratios in stdlib made me assume our serializer was good enough to ignore, but that is now looking like clearly a mistake.)

Signed-off-by: Eric Myhre <hash@exultant.us>