Allocs 14->9.  Put ptrderef machines into slab.
This gets our struct benchmark almost back in line with where it was pre-slab, but still isn't a concrete win, and arrays are still seeing some punishment.  Bit surprised.

Amend: confirmed by looking at cpu profile graphs: 'runtime.mallocgc', the function that indirects your allocations while the gc is running, is now a pretty huge fraction of the time spent for the array benchmarks.  So, yes: sheer increase in memory pressure from the slab technique is penalizing here, and we're benching the wrong thing by not reusing marshallers.

Signed-off-by: Eric Myhre <hash@exultant.us>