Flip benchmarks to reusing objects.
Do this for both the stdlib stuff, and ours, so the comparison remains apples to apples.

This reduces the stdlib mallocs all the way down to one -- for everything.  Ours drop, though of course not by that much (yet -- the payoff combining this and the slabs should be nonlinearly awesome).

Stdlib gets a fair amount of CPU back, in no smaller part just due to buffer reuse that's also a part of this change (which negates the 'grow' costs); so do we.

It's possible stdlib is actually getting a slight penalty now because it may not know, internally, that it's writing to a bytes.Buffer, and may be hitting more string->[]byte conversion costs as a result of lacking the WriteString method.  However, we've been playing with that same handicap this entire time, so that also seems not unfair.

Signed-off-by: Eric Myhre <hash@exultant.us>