Extract token sequences from json tests to fixture package.
These should be reusable -- certainly for other serializers and deserializers, at least.  The object mappers might be another story because they need to explore quite a few other parameters, but we can aim for as much reuse as possible there too.

One could also imagine building one megastructure of all facets of the fixtures, where we put the cbor bytes and the json strings and the tokens and so forth all in one struct, and check the N-way compatability in the relevant packages.  We're not, and instead taking this approach of token sequences and mapping the package's real test list onto them, because that megastruct approach really only makes sense if you make the mistake of believing everything is going to be 1-to-1.  Sure, sometimes it is, and that's the goal; but many (many) of our formats tolerate non-canonical encodings, and our test structures certainly shouldn't make that *less* easy to clearly label and text matrices of.

Signed-off-by: Eric Myhre <hash@exultant.us>