Add some corrections to the copy-by-* microbenchmark.  What we're measuring there is mostly 'runtime.convT2E'.
Previous postulations in comments about strings were incorrect.

Now, in context: mind that the distinction between copy-by-value on a *structure* and return-a-pointer is a whole different ballgame.  But what reasoning we apply to the way we pass token information back and forth vs what kind of techniques we want atlases to support during object mapping are also whole different ballgames; this benchmark's info is primarily applicable to the token-passing game, and there ARE other alternatives we can entertain here (such as type-slabs to remove the need for interface wildcards entirely, since we DO have a finite set of valid types of token content).

Signed-off-by: Eric Myhre <hash@exultant.us>