A wild cbor decoder appears!
The terminals are largely with the help of some code previously written by github.com/ugorji/go/codec -- also fortunately under MIT license terms, and so free and compatible with ours in this project.  The state machine stuff is all new code, since of course our token-yielding architecture is just fundamentally different.  By comparison to that library, this leaves us with a lot of interesting simplifications (or, complexities-pushed-elsewhere): we always yield uints for positive integers for example, because we have no other information (e.g. a type we know we're mapping into) to suggest we do otherwise.

The reader helpers are also ugorji-inspired, though I've already made some tweaks to fix some really weird crosscutting knowledge his callers had about when zero-copy is happening, and also expect to refactor more in the future to handle errors a little more... so?

Tests refer heavily to the token fixtures we extracted from the json package this morning, so that paid off nicely already.  Coverage is currently missing for many other kinds of literal terminals, but things essentially all seem to be working.

Handling for definite-length arrays and maps currently absent, which is a big oversight -- that's coming up momentarily.  Support for tags also currently absent -- that one's going to take a little more doing, because it fundamentally requires more information than the tok.Token currently conveys.  (Spec'ing token fixtures for input into definite-length *encoding* tests in the future will *also* require more information out of the tok.Token, coincidentally!)  So we'll have a merry (no-so-)little refactor to tokens coming up shortly.  Hopefully all of the existing code can be migrated with sed.

Signed-off-by: Eric Myhre <hash@exultant.us>