Move fixed-len map termination to key step.  Simpler this way.
Temptation was to have that endMap step as its own thing so we could jump straight to it and skip pushing a "0" state for empty arrays... but that's an append on an array that spends most of its time having capacity already; it's inconsequential.  And going that way puts the zero-entries check in *two* places.

Signed-off-by: Eric Myhre <hash@exultant.us>