Reshuffle some fixtures; comments on groupings.
And comments on how dreamy it would be if anyone could agree on *what numbers are*.  It's kind of amazing what a pandemic this is.  I think people are often peripherially aware of this, but wow does it really burn to look into the naked vortex at the heart of a system like this where you dare to try to map two concepts together.  It's *fortunate* in a way that CBOR and JSON both have sufficiently overlapping(-ish) ambiguity zones that they tend not to cause each other nightmares.  But this really lays bare when it's so popular to build either external schemas with codegen, or, for the same reasons, rely on object-mappers, with no direct ability to stream from serial format to serial format: It's downright *harrowing* to try to "do the right thing" for numerics without hitting an edge case on one side or the other.

Given the above, it's worth making a positive note: in general, this library still has utility, because if you map into some other schematized format at any point, you're still playing by precisely that good of ruleset.  If bouncing something through a schematized format, with more specifically bounded casting configurations, improves your data handling: well then, cest la vie.

Signed-off-by: Eric Myhre <hash@exultant.us>