Support for floats; fix an initialization step and some dire bugs flushing larger ints; tests for floats and larger ints.
Float32 will not currently be supported.  If we do, it will probably be by adding a TFloat32 to the enum of TokenType.  It might be nice to automatically detect when numbers can be losslessly casted to a smaller float encoding, but that road is fraught with peril, and the CBOR spec does not seem to demand it (remarkably, even in the canonicalization section; it seems they thought (and I agree heartily, fwiw) that if you're concerned with canonical representations of data at all, you simply *shouldn't be using floats* at all).

Signed-off-by: Eric Myhre <hash@exultant.us>