Add benchmarks for cbor marshal, paralleling existing benchmarks covering json.
CBOR is faster than JSON -- no big surprise -- but not by a huge amount on encodings.  (We'll expect a bigger gap on *decoding*, since that's where the fixed-length stuff really pays off.)

We still have performance improvements to make in the object-mapping layers themselves to get where we really want to be.

There's one unpleasant mark here: the struct marshalling benchmark is *worse* for cbor.  Why?  Well, the 25 allocations is a hint...

Signed-off-by: Eric Myhre <hash@exultant.us>