Improve performance of cbor encoder when writing strings to a writer than can handle them efficiently.
If your output stream has a `WriteString` method, this will dramaticaly reduce your number of allocations.

Future additional options for improvements here include unrolling the string into a scratch byte slice (which the compiler should likely be smart enough to unroll into a duff's device, or similarly performant choice -- certainly something without a bunch of unnecessary stack growth and also without allocs, anyway), and calling chunked writes of the size of the scratch buffer; or, making use of `unsafe` to get a direct view of the bytes under the string (but would be somewhat alarming to hand to an arbitrary writer functon...).

Signed-off-by: Eric Myhre <hash@exultant.us>