Begin draft of atlas with more comprehension of non-struct kinds.
This was hinted at in the previous generation with the Atlas/Morphism split but A) it wasn't really realized and B) both those names in flight at the same time is extremely confusing.

The previous generation was also deeply burdened by the idea that all the structs would make sense to lay out declaratively with public fields all the way.  They don't.  There are a *lot* of initializers and sanity checkers and precomputations in this system which make purely declarative structs very unpleasant.

Signed-off-by: Eric Myhre <hash@exultant.us>