Begin porting over marshaller code in bulk, but filtered to use reflection consistently (giving up the hideously complex addrfunc concepts entirely).
The addrfunc idea is neat, but utterly nightmare-inducing to debug, and supporting both that and the (much more common) reflect path at the same time pretty much squares the complexity of the code, with corresponding (if not quite so supralinear) degradations to runtime as it switches normalize forms back and forth.  Let's just see how far we get without it.

Signed-off-by: Eric Myhre <hash@exultant.us>