Implement test look for marshaller.
Skip it out because we still have enough things stubbed with nil returns that panics are instantaneous.

Remove previous references to expectPanic.  Screw that.  The expectation is "no panics".  (Mostly this is because it's just plain debugging hell when it comes to comprehending what the heck is up when you use a naive func like our testutil.CapturePanics here and it discards all the stack info.  If we were importing meep errors and using that kind of stackkeeping technology?  Maybe it would be a different story.)

Signed-off-by: Eric Myhre <hash@exultant.us>