Implement primitive marshalmachine.
This is actually way easier now that we're going full reflection.  It turns out the accessors on reflect.Value pretty much mirror exactly what I was normalizing things to in the previous generation of code anyway -- it fits straight on to the TokenType enum, even all the way down to details like uintptr being normalized to a uint64 by the Value.Uint method.

Signed-off-by: Eric Myhre <hash@exultant.us>