Big bag 'o unmarshaller.
This code is huge amounts of mirroring of the marshaller stuff.  The machines at the end are all different, of course.

Starting to use some typed errors, since our test fixtures we laid out well in advance are starting to drag us into testing error paths right off the bat.

Signed-off-by: Eric Myhre <hash@exultant.us>