Upgrade TestMarshaller to use new test framework.  huge readability boost.
Also, if we had a huge bunch of test table entries fanning out, some of which panicked, whatever -- like the unmarshaller tests already do, and as the marshaller tests will begin to as we write more of them! -- oh MAN is this a useful bulwark against a single panic causing an untold percent of your entire suite to not even bother running.

(I really love this library.  I should've brought it in ages ago.)

Signed-off-by: Eric Myhre <hash@exultant.us>