Upgrade TestUnmarshaller to use new test framework.  I can see for miles and miles...
Seriously:

```
  Marshaller suite:
    "string literal" fixture sequence:
      Steps... ✔
        Result (halted correctly) ✔

  Unmarshaller suite:
    "string literal" fixture sequence:
      targetting into string:
        Result (error expected) ✔
      targetting into *string:
        Steps... ✔
          Result ✔
      targetting into wildcard:
        Result (error expected) ✘
      targetting into *wildcard: 🔥
      targetting into map[str]iface:
        Result (error expected) ✘
      targetting into *map[str]iface: 🔥
      targetting into []iface:
        Result (error expected) ✘
      targetting into *[]iface: 🔥
```

Ain't that NICE?!  Okay, the 'fire' emoji gets mangled when I paste into vim.  But now we get a bird's eye view of the whole matrix of what's good and what's fucked, and all the panics (yes, all of them, because they're managed correctly within a tree now) are listed after this summary.  Beauty.

Signed-off-by: Eric Myhre <hash@exultant.us>