Admit sanity test about semantics of reflective map set.
There was a moment where I mis-parsed the godocs for this function -- "zero" is a somewhat overloaded word in that context.  But indeed, everything adds up -- it's the zero reflect.Value (not the "zero value" for the type); a.k.a. a value with Kind==Invalid, which causes deletion.  Adding a zero value of a real type is as you'd expect.

Signed-off-by: Eric Myhre <hash@exultant.us>