Ah, this is interesting: we *can* unmarshal into a map... as long as it's already been made.
It's the initialization that's the only source of concern for maps: You'll get a "reflect: reflect.Value.Set using unaddressable value" panic if starting from a nil, un-pointer map.  Starting from the pointer means we CanAddr it and thus can initialize it.  You can set keys into a map even if you don't have CanAddr to it.

The updated gating logic in UnmarshalDriver.Bind appears to accomodate this correctly, though I'm not sure if it's the most succint or semantic way to express it.

Signed-off-by: Eric Myhre <hash@exultant.us>