Support and test for maps of other value types.
Required adding an `Elem()` call to peek through the `interface{}` slot and get the correct type, so the checks in SetMapIndex wouldn't reject it.

Not sure if there are other, better ways to implement this (perhaps such as `New`'ing the right type, then using that?  But would that require *more* mallocs overall since it would be that much harder to recycle?  Or is t2conv hitting us anyway?)... benchmarking will be required to answer any of these questions, and we'll return to this later after getting more feature completeness.

Signed-off-by: Eric Myhre <hash@exultant.us>