Add test for marshalling from pointers to maps!  One quick fix.
Made equivalent fix on the unmarshal side, which had the mirrored error, though so far there it has been innocuous.

This shows up where it does because we have so far ended up mostly *not* using the reflect.Type value in Reset methods.  Only maps and arrays are likely to use it, for the most part.  And the unmarshalMap machine is currently a bit more shoddy than the marshal one, and recomputes the type value far more than necessary, so was coincidentally immunized from this.

Signed-off-by: Eric Myhre <hash@exultant.us>