Forward on working with slice unmarshal.
Also test coverage for marshalling non-empty slices, but that was pretty much expected to go off without a hitch.

Works on paths where you have an addressable handle to a slice (whether the precise type, or `[]interface{}`).  Missing a bit when the wildcard machine is in action: we need a post-slice-unmarshalling step where the final slice is placed into its spot -- there's no way to fill it in sensibly as we go (as far as I've been able to hash out yet, anyway; the only visible options include different depths of 'Elem()'ing per whether or not you're boxed in an interface{} slot, which is untenable).

Signed-off-by: Eric Myhre <hash@exultant.us>