Test coverage for complex nested structure.  Passes, woo!
This is *still* *not enough* coverage for all the myriad little twisty passages that reflection could potentially need to follow, though.  Specifically, we actually got an interesting free pass here: because it's a map with `interface{}` value, we actually hit the wildcard machine and are just giving that indirection a *ton* of exercise.

Signed-off-by: Eric Myhre <hash@exultant.us>