Test using atlas with multiple types; tweak some existing token fixtures to be reusable for that; fix major issue with addressing in a for-loop causing multi-entry atlases to be initialized incorrectly!
Builders for atlas configuration shall now return *AtlasEntry pointers instead of the structs directly.  Can you spot why?  That goddamn landmine of go reference-taking semantics bit me: atlasBuilding.go:12 in collaboration with line 17.  I understand exactly why that works the way it does -- but I can't help but feel this is the sort of shitty landmine that, unfocusing your eye for a minute and thinking about what programming languages are supposed to save you from, this is just precisely the sort of problem I want a strongly typed language to immunize me from.

Anyway.  Builders for atlas configuration show now return *AtlasEntry pointers.  The pointer form was mostly being used internally already anyway.  All this does is give the user an opportunity to footgun themselves by holding on to the pointer after building the atlas and mutate it later in unacceptable ways.  So, under the "we're all consenting adults here" paradigm, don't do that.

Signed-off-by: Eric Myhre <hash@exultant.us>