Support struct unmarshal, and test coverage.
Somewhat shockingly, the code for unmarshalMachineStructAtlas is *exactly* the same as the mirrored marshal machine.  The only distinctions are the slab and driver types that are handed down.  This makes sense: these machines do very little to the struct itself; they're almost entirely in the business of iterating fields and dispatching recursion.

Signed-off-by: Eric Myhre <hash@exultant.us>