Rename json s/serializer/encoder/g.
This is a consistency fix to match the cbor packages.

Generally: we're going to use the words encoder and decoder for things that exchange a token stream for bytes; marshaller and unmarshaller for things that map objects and memory to token streams; and serializer and deserializer are words that... are a little unclear: they refer to any part that's linearizing things, I suppose, but that tends to be the marshaller/unmarshaller, which makes the word redundant; we'll probably mostly avoid it, or perhaps use the word "serialization" refer to the whole picture (e.g. when you assemble a marshaller and an encoder, now you've got a serializer).

Signed-off-by: Eric Myhre <hash@exultant.us>