Begin work on transformers: funcs that enable transforming any value into another... even of other kinds!
It's going to be a large swatch of work to get the min viable thing out, so test skipping added here, and errors still punting fiercely.

There's already indications we might need to reconsider our function types a bit and make sure we keep all the type data in tuples, because otherwise incorrect assembly will be entirely too easy and checking for it entirely too hard.

Signed-off-by: Eric Myhre <hash@exultant.us>