Test filling nulls back into strings, string pointers, and string pointer-pointers.  (Only the last is valid.)
I'm not sure that second behavior is correct.  It seems reasonable if you take it as read that the unmarshaller only receives a pointer in the first place so it can set things of the dereference type, and a null is not a valid value for a string.  But that would *not* be what you wanted if you got a null for a field of type `*string` in a struct.  Fortunately, we'll be testing exactly that in a bit (see we already added the token fixtures for it at the beginning of this branch?), so we can review then.

Signed-off-by: Eric Myhre <hash@exultant.us>