Test unmarshalling nils into maps.  Requires a significant fix.
We were wrong back in 748ebd65cf0831fc606ded678a425ff6f0f1e66d: there's a reason to reject maps in unmarshal.  Even if you can set _into_ them, you also still need to be able to set the map itself, so if it's being reused, and you get a null token, you can properly set the map to nil!

This required changing a bunch of previous test fixtures to now treat this as an error.

Signed-off-by: Eric Myhre <hash@exultant.us>