rename: strip "Driver" from obj.Marshaller and obj.Unmarshaller names.
The "driver" part may have made more sense if they were unexported types, but since we've seen no major reason to make them interfaces and hide the concrete type, we might as well have the shorter names.  Several docs already allude to things in that way.  This is also following the conventions of the stdlib json package's type names.

Signed-off-by: Eric Myhre <hash@exultant.us>