Add json decoder!
The state machinery is largely a riff of the cbor decoder -- as it should be -- plus some stuff for munching whitespace and handling colons and commas.

Handling terminal leaf types todo; this is just all the big stuff and recursive parts.  (String escape handling and float stuff in json is actually pretty dang nontrivial, so we're going to give those their own commits.)

Basic tests pass; many more to come.

Signed-off-by: Eric Myhre <hash@exultant.us>