Replacing and reorganizing the porcelain layers!
We're going to attempt to have package methods that look a little more familiar to folks coming from the golang stdlib... like this:

```
json.NewMarshaller
json.NewAtlasMarshaller
obj.NewTransformer
refmt.NewMarshaller(json.Config{}, ...)
```
This will take several commits.  In this one, just cbor getting started:

Move helper functions for assembling cbor<->obj and json<->obj into their respective serial packages.

The upside of this is the calls -- e.g. `json.Marshal` -- look more like what is familiar from the standard library.

The downside is that all the serial packages now import the object package.  This bothers me from a purist perspective, because the json package certainly doesn't *need* to have a dependency tree which includes the `reflect` package -- but probably isn't a problem in practice because the number of use-cases in the wild that *don't* involve the obj package anyway are very nearly zero.  (Even the refmt cli tools will end up with it in the binary.  It doesn't matter if a user only uses a cbor->json transformer command or the like; we're certainly not going to be bothering to ship special binaries to shave a meg off disk for that use case.)

We'll also be adding more generalized functions back to the top level package shortly, which you will be able to swap the behaviors of by using a config object typethunk.

Some types like TokenPump are now currently duplicated; we'll be aiming to remove the one from the top-level refmt package by the end of this.  (It both *can't* fit there anymore for import cycle reasons, but also was pretty ugly to have there in the first place; that low, low, LOW level implementation detail would really preferably not be on the first page of godoc a user sees.)

Signed-off-by: Eric Myhre <hash@exultant.us>