Fix spelling of Marshalling.
I can't take it anymore.  The golang stdlib spelling is just *wrong*.  I see the historical argument.  I don't care.  That's not how english works and I'm not propagating this bizarrotron exception to the way any sensible adult would expect this word to be spelled.  Other programming languages with marshallers spell it "marshaller".  The dominant spelling by the sheer heuristic of google matches is "marshaller", by far.  Search for "marshaling", and you'll get "Did you mean: marshalling".  The wikipedia page is "Marshalling_(computer_science)".  Enough.  The agreement is *resounding*.

My fingers have twitched and cramped from the struggle to resist that second 'l' for long enough.

Sed: attack.

(The real straw that broke the camel's back, and in this commit on this branch in particular, is the realization that we can't easily brook the stdlib 'encoding/json' function names anyway, because what we call an "encoder" specifically means it deals with bytes and tokens; and the 'encoding/json' package does not share this opinion.  So, migrating to refmt is going to take you some minimal amount of sed instead of mere import rewriting *anyway*.  In that light, I'm out of even the flimsiest reasons not to make this spelling correction.)

Signed-off-by: Eric Myhre <hash@exultant.us>